stored byte[][] from files into map
