+++ application-server/conf/application.properties
@@ -16,3 +16,4 @@ system.log.level=DEBUG
 core.site.conf.location=/s/bach/g/under/evanjs/Documents/Senior/CS555/hadoop-conf/core-site.xml
 hdfs.site.conf.location=/s/bach/g/under/evanjs/Documents/Senior/CS555/hadoop-conf/hdfs-site.xml
 
+hdfs.file.location=/sectors/sector.txt
